Add IMU interface that's implemented by Pigeon classes
With the addition of the Pigeon2, we have two versions of the same
sensor in use and available to us. This abstracts the details of
which IMU is installed on the robot such that we can easily swap
it out in or out on the hardware side while requiring minimal
changes on the software side of things.