Refactor robot configuration into interface
As I'm been working on doing other configuration work, I've stumbled
into some of the cumbersome sides of the all-static setup. Specifically
there's no way to ensure that a robot config has specific fields, and
you also have to use the conditional-stack we had in Swerve.java in
order to handle all the robots.

This change refactors all of the configs into a common interface, and
then uses that to consolidate the creation of our Swerve. Now adding
a new robot only requires declaring the config and changing the name,
and adding a new field to the robot config doesn't require changes in
several different places (once per robot).