Move configuration and constants to their own directories
This is part of some of the larger cleanup efforts
to make the code/folders more parsable. In these
instances the classes moved here were either within
or alongside subsystems which made them harder to find
and added extra code in those classes.

This is intended to be a pure configuration change with
no effect on functionality.