Apply VS Code formatters to all of the cyberlib files
This change is purely stylistic and makes 0 code changes.

The intent behind this is to ensure a consistent style no
matter where the code is being viewed or who wrote it. To
that end I've applied the following changes to everything
under the cyberlib package

1. Set document tab size to 4 spaces
2. Run VS Code "Convert Indentation to Spaces"
3. Run VS Code "Organize Imports"
4. Run VS Code "Format Document"

This can be run on any system so that anyone can format
their code in line with what's been set here.