Fix driver flow for resetting the field-oriented drive
Previously we have called Swerve.zeroSensors() which
both sets the robot position and updates the IMU. This
has caused issues which incidentally was fixed by calling
zeroSensors twice, but that's not the fix we _should_ be
doing.

This updates the code to instead just call setRobotPosition,
which handles things as expected, and adds a comment to avoid
this from happening again in the future.