Remove SwerveModuleConstants class
This was intended to happen when I created the SwerveModuleConfiguration
class in #38, but I apparently forgot to remove the original
class and update it's usages. This completes the refactor
started there.

I verified when doing this that the values were still the same
between SwerveDriveConstants and SwerveDriveConfiguration so
this should be a no-op. But if it proves not to be feel free to
revert this.