Update most sensor reads to all happen in one spot
We follow this pattern in most of the code, but
there were a few spots where we're doing direct
reads from the motors outside of readPeriodicInputs().
This change cleans up a lot of those interactions,
adding fields in mPeriodicIO where needed.

The only difference is we read stator currents
directly in telemetry to avoid CTRE errors in the
rio log.