Add independent tracking of mid arm and slappy in Stage3
This was proposed last night but not implemented because
it didn't yet become a problem, but thinking on things
more (both last night and today) I think it's important
to add.

In stage 3 we move both the slappy sticks and the mid
arm at the same time, and then wait for both to come into
place before moving forward. Instead of requiring they
both do that at the same time, this adds independent
tracking so that once either of them is in place once
we set that flag and move forward.