Add code to create a slow config for auto trajectories
Auto paths are generated based on the TrajectoryConfig object, which
inherits the robots max speed. We don't want to change the max speed
for all autos, but need a faster than normal max speed for the
5-ball auto.

This change allows us to keep the "slow auto config" for the existing
autos while independently changing the speed for the fast autos in a
separate commit to come later.